separated pca code
